third_party/WebKit/LayoutTests/fast/frames/sandboxed-iframe-navigation-top-by-name-denied.html

<html>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
    var test = async_test("Sandboxed frame cannot navigate top-level browsing context by name.");

    window.addEventListener("message", test.step_func(function (e) {
        assert_equals(null, e.data["target"]);
        assert_equals(null, e.data["_top"]);
        assert_equals(null, e.data["_parent"]);
        assert_equals(null, e.data["_blank"]);

        test.done();
    }));
</script>
</head>
<body>
<iframe sandbox="allow-scripts"
        srcdoc="
<script>
  window.top.postMessage({
    'target': window.open('fail-and-notify-done.html', 'target'),
    '_top': window.open('fail-and-notify-done.html', '_top'),
    '_parent': window.open('fail-and-notify-done.html', '_parent'),
    '_blank': window.open('fail-and-notify-done.html', '_blank'),
  }, '*');
</script>
        ">
</body>
</html>
